feat: Add inquisitor highlighting feature
This commit adds a new feature to highlight inquisitors in the game. The code changes include registering a new render function to draw a colored box around inquisitor entities. The color of the highlight can be customized using the new "Inquis Color" setting in the Diana category of the settings. The highlight is only applied if there are inquisitor waypoints and the "Highlight Inquis" setting is enabled.