feat: Add support for additional party commands
This commit adds support for additional party commands "!chims", "!chimeras", "!book", and "!books". When these commands are used and the "dianaTracker" setting is enabled, the code will send a command to the server to retrieve the corresponding item count from the "dianaTrackerMayor" object. This enhancement expands the functionality of the party commands feature, allowing users to retrieve item counts for more commands.