refactor: Enable Diana Party Commands for sending counts in party chat
Add conditional checks in PartyCommands.js to enable the Diana Party Commands feature for sending counts in party chat. This ensures that the commands for chimera, inquisitor, relic, stick, mob, burrow, and since will only be executed if the Diana Party Commands setting is enabled. This enhancement improves the functionality and flexibility of the party commands feature.