refactor: Update DianaGuis.js to improve loot message formatting
This commit updates the getMessagePart function in DianaGuis.js to improve the formatting of the loot message. It adds a new parameter "type" to differentiate between different types of loot items. Specifically, it adds a special formatting for the "Chimera" item, displaying the percentage and the number of "ChimeraLs" in the message. This change enhances the readability and clarity of the loot message in the Diana Loot Tracker.