refactor: Update PartyCommands.js to display lootshare count in chat command
This code change updates the "register" function in PartyCommands.js to display the lootshare count in the chat command for the "Chimera" item. The lootshare count is calculated using the "dianaTrackerMayor" object and the "calcPercentOne" function. The chat command now includes the lootshare count in addition to the regular Chimera count.