refactor: Update "Diana/PartyFinder.js" and "utils/checkDiana.js" to fix party checking logic
This code change updates the "Diana/PartyFinder.js" file and "utils/checkDiana.js" file to fix the party checking logic. In "Diana/PartyFinder.js", the "checkDiana" function is now called with the parameter "true" to override the check for the "Hub" world. In "utils/checkDiana.js", the "checkDiana" function now accepts an "overrideHub" parameter to control the check for the "Hub" world. These changes ensure that the party checking is done correctly and improve the reliability of the code.

Co-authored-by: [Author Name]