refactor: Update PartyFinderGUI to use new API endpoint for getting all parties
This code change in PartyFinderGUI.js updates the getPartyFinderData function to use the new API endpoint "https://api.skyblockoverhaul.com/getAllParties" for retrieving all parties. It replaces the previous implementation that used the getAllParties function from the Diana/PartyFinder module. This refactor improves the reliability and performance of the PartyFinder feature.

Co-authored-by: [Author Name]