refactor: Update DianaMobDetect.js and DianaTracker.js
This refactor updates the DianaMobDetect.js file to include an additional parameter "skytils" in the register function for chat events. It also updates the DianaTracker.js file to include an additional parameter "skytils" in the register function for chat events. These changes improve the functionality and performance of the DianaMobDetect and DianaTracker features.