feat: Add deQueueButton to PartyFinderGUI.js
This commit adds the deQueueButton to the PartyFinderGUI.js file. The deQueueButton is a new button that allows users to dequeue from a party. When clicked, it triggers the "Dequeueing" message to be sent in the chat. This change enhances the functionality of the PartyFinderGUI by providing a way for users to easily remove themselves from a party.

Co-authored-by: [Author Name]